% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{extdata}
\alias{extdata}
\title{FoxA1 ChIP-exo experiment}
\format{
bam files (with their respective indexes)
}
\description{
bam file with the reads aligned to chr1 from FoxA1 in mm9 liver cell lines
published by Serandour et al., 2013. This package contains the files
}
\details{
\itemize{
\item ChIPexo_carroll_FoxA1_mouse_rep1_chr1.bam
\item ChIPexo_carroll_FoxA1_mouse_rep2_chr1.bam
\item ChIPexo_carroll_FoxA1_mouse_rep3_chr1.bam
}
and their respective index files.
}
